web/MyNotes/deploy/app/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy==2.0.30
Werkzeug==3.0.3
